requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
sentence-transformers==3.2.1
faiss-cpu==1.8.0.post1
tqdm==4.66.5
readability-lxml==0.8.1
regex==2024.11.6
streamlit==1.38.0
